requirements.txt
skir-client >= 0.0.2
flask >= 3.1.0
